chore: finish restructure and fix compiler
